+++ packages/@cdk-cosmos/core/src/features/index.ts
@@ -1,3 +1,5 @@
+export { ILinkFeature, LinkFeatureStack } from './link-feature/link-feature-stack';
+
 export {
   ISharedVpcFeature,
   SharedVpcFeatureCoreStack,
@@ -9,7 +11,6 @@ export {
   EcsFeatureCoreStack,
   EcsSolarSystemCoreStackProps,
 } from './ecs-feature/ecs-feature-core-stack';
-
 export { EcsFeatureCoreImport, EcsFeatureCoreImportProps } from './ecs-feature/ecs-feature-core-import';
 
 export {
@@ -17,9 +18,24 @@ export {
   CiCdFeatureCoreStack,
   CiCdFeatureCoreStackProps,
 } from './cicd-feature/cicd-feature-core-stack';
-
 export {
   ICiCdFeatureExtension,
   CiCdFeatureExtensionStack,
   CiCdFeatureExtensionStackProps,
 } from './cicd-feature/cicd-feature-extension-stack';
+
+export { Domain, DomainProps } from './domain-feature/domain-feature-stack';
+export { Subdomain, SubdomainProps } from './domain-feature/subdomain-feature-stack';
+
+export { IRedisFeature, RedisFeatureStack, RedisFeatureStackProps } from './redis-feature/redis-feature-stack';
+export { RedisFeatureImport, RedisFeatureImportProps } from './redis-feature/redis-feature-import';
+
+export {
+  IGithubEnterpriseFeatureCore,
+  GithubEnterpriseFeatureCoreStack,
+  GithubEnterpriseFeatureCoreStackProps,
+} from './github-enterprise-feature/github-enterprise-feature-core-stack';
+export {
+  GithubEnterpriseFeatureCoreImport,
+  GithubEnterpriseFeatureCoreImportProps,
+} from './github-enterprise-feature/github-enterprise-feature-core-import';